refactor(table): rename editing directory to edit
