+++ js/app/src/components/table/edit/EditableJSONCell.tsx
