+++ js/app/src/components/table/edit/jsonEditorValue.ts
